agent crud done

+++ src/agents/agents.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { AgentDto } from './dto';
+import { AgentDto } from './dto/agents.dto';
 import { Prisma, PrismaClient, Agent } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { AgentEntity } from './entities/agents.entity';
@@ -10,15 +10,28 @@ export class AgentsService {
 
   async create({ agentDto }: { agentDto: AgentDto }): Promise<AgentEntity> {
     try {
-      const agentWithEmail = this.prisma.agent.findUnique({
+      // find any agent with the provided email
+      const agentWithEmail = await this.prisma.agent.findUnique({
         where: { email: agentDto.email },
       });
 
+      // throw an error if an agent is found
       if (agentWithEmail) {
         throw new Error('Email already used');
       }
 
-      return await this.prisma.agent.create({
+      // find any agent with the provided phone number
+      const agentWithPhoneNumber = await this.prisma.agent.findUnique({
+        where: { phoneNumber: agentDto.phoneNumber },
+      });
+
+      // throw an error if an agent is found
+      if (agentWithPhoneNumber) {
+        throw new Error('Phone number already used');
+      }
+
+      // create a new agent
+      return this.prisma.agent.create({
         data: agentDto,
       });
     } catch (error) {
@@ -49,6 +62,7 @@ export class AgentsService {
     orderBy?: Prisma.AgentOrderByWithRelationInput;
   }): Promise<AgentEntity[]> {
     try {
+      // fetch all agents with the specified parameters
       return await this.prisma.agent.findMany({
         skip,
         take,
@@ -77,19 +91,19 @@ export class AgentsService {
 
   async findOne({ id }: { id: number }) {
     try {
+      // fetch agent with the provided ID
       const agent = await this.prisma.agent.findUnique({
         where: { id },
       });
+
+      // throw an error if any agent is found
       if (!agent) {
-        throw new Error(`Entity with ID ${id} not found`);
+        throw new Error(`Agent with ID ${id} not found`);
       }
+
+      // return the requested agent
       return agent;
     } catch (error) {
-      if (error instanceof Prisma.PrismaClientKnownRequestError) {
-        if (error.code === 'P2025') {
-          throw new Error('Record not found');
-        }
-      }
       if (error instanceof Prisma.PrismaClientUnknownRequestError) {
         throw new Error('Invalid query or request');
       }
@@ -103,24 +117,51 @@ export class AgentsService {
     }
   }
 
-  async update({ id, agent }: { id: number; agent: AgentDto }): Promise<Agent> {
+  async update({
+    id,
+    agentDto,
+  }: {
+    id: number;
+    agentDto: AgentDto;
+  }): Promise<Agent> {
+    console.log({ agentDto: agentDto });
     try {
+      // fetch agent with the provided ID
+      const agentWithID = await this.prisma.agent.findUnique({
+        where: { id },
+      });
+
+      // throw an error if any agent is found
+      if (!agentWithID) {
+        throw new Error(`Agent with ID ${id} not found`);
+      }
+
+      // find any agent with the provided email
+      const agentWithEmail = await this.prisma.agent.findUnique({
+        where: { email: agentDto.email },
+      });
+
+      // throw an error if an agent is found and it is not the requested agent
+      if (agentWithEmail && agentWithEmail.id != id) {
+        throw new Error('Email already used');
+      }
+
+      // find any agent with the provided phone number
+      const agentWithPhoneNumber = await this.prisma.agent.findUnique({
+        where: { phoneNumber: agentDto.phoneNumber },
+      });
+
+      // throw an error if an agent is found and it is not the requested agent
+      if (agentWithPhoneNumber && agentWithPhoneNumber.id != id) {
+        throw new Error('Phone number already used');
+      }
+
+      // update the agent data
       return await this.prisma.agent.update({
         where: { id },
-        data: agent,
+        data: agentDto,
       });
     } catch (error) {
-      if (error instanceof Prisma.PrismaClientKnownRequestError) {
-        if (error.code === 'P2002') {
-          throw new Error('Unique constraint violation');
-        }
-        /* if (error.code === 'P2003') {
-          throw new Error('Foreign key constraint violation');
-        }*/
-        if (error.code === 'P2025') {
-          throw new Error('Record not found');
-        }
-      }
       if (error instanceof Prisma.PrismaClientUnknownRequestError) {
         throw new Error('Invalid query or request');
       }
@@ -136,17 +177,22 @@ export class AgentsService {
 
   async remove({ id }: { id: number }): Promise<AgentEntity> {
     try {
-      const agent = await this.prisma.agent.delete({ where: { id } });
-      if (!agent) {
-        throw new Error(`Record with ID ${id} not found`);
+      // fetch agent with the provided ID
+      const agentWithID = await this.prisma.agent.findUnique({
+        where: { id },
+      });
+
+      // throw an error if any agent is found
+      if (!agentWithID) {
+        throw new Error(`Agent with ID ${id} not found`);
       }
+
+      // remove the specified agent
+      const agent = await this.prisma.agent.delete({ where: { id } });
+
+      // return removed agent
       return agent;
     } catch (error) {
-      if (error instanceof Prisma.PrismaClientKnownRequestError) {
-        if (error.code === 'P2025') {
-          throw new Error('Record not found');
-        }
-      }
       if (error instanceof Prisma.PrismaClientUnknownRequestError) {
         throw new Error('Invalid query or request');
       }
